+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Handle client certificate expiration by requesting new credentials to Pairing API.
+
 ## [0.10.2] - 2019-12-09
 ### Added
 - Allow setting an explicit hwid instead of using the hardware generated one.